start of what will be 2.0 base
sipwitch will be transitioned to the new voip intermediary
library which mediates between eXosip2 3.x and the new
eXosip2 4.0 api, the latter of which supports multiple sip
protocol contexts.  This checkin only represents the starting
point of this process.

It is possible there will be intermediary release(s) produced
between 1.5 and 2.0 which will have transitional code,
particularly if a critical bug is found in 1.5, or a new
feature is deamed essential before 2.0 is ready.  However,
the goal is to go directly from 1.5 to 2.0.

One result of this is that we will be able to continue building
and using sipwitch on gnu/linux distributions that still contain
eXosip2 releases prior to 4.0, albiet without the ability to
support concurrent protocols.  Since many long term supported
distributions have released recently with eXosip2 3.x libs, we
expect such support to continue to be used for a long time.