restaging events because some platforms are broken
The problem is that some systems (openbsd, osx, etc) will deadlock
on an accept call even if the underlying socket has been closed.

This means we will have to create a dummy connection in event
terminate to get past the blocked accept call with a shutdown
flag set.  This is staging for that work, as it also has to be
done differently for broken windows, too...